requirements.txt
# Core
gradio>=4.31,<5.0
pydantic>=2.5
pyyaml>=6.0
numpy>=1.24
opencv-python-headless>=4.9
Pillow>=10.0

# Torch stack (pin as needed)
torch
torchvision

# Optional extras used in prior pipeline
mediapipe; platform_system != "Windows"
timm
matplotlib
